AKI-414: Convert chargeEnergy helpers from long to int
-while our worst-case code instrumentation growth size is still 6 bytes (invokestatic + ldc_w - even if ldc is the common case), it is still preferable to favour the creation of smaller transformed code
-additionally, this means that we have a chance to impose strict limits on non-constant billing values, verifying that they are within the domain of positive integers

+++ org.aion.avm.core/test/org/aion/avm/core/MockFailureInstrumentationFactory.java
@@ -49,7 +49,7 @@ public class MockFailureInstrumentationFactory implements IInstrumentationFactor
                 return underlying.wrapAsThrowable(arg);
             }
             @Override
-            public void chargeEnergy(long cost) throws OutOfEnergyException {
+            public void chargeEnergy(int cost) throws OutOfEnergyException {
                 int thisIndex = MockFailureInstrumentationFactory.this.count.getAndIncrement();
                 if (thisIndex == MockFailureInstrumentationFactory.this.basicBlockIndex) {
                     MockFailureInstrumentationFactory.this.toRun.run();
